othello/requirements.txt
click==8.1.3
fastapi==0.115.5
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.21.6
tensorflow-cpu==2.9.1
Werkzeug==2.2.2
zipp==3.10.0
gunicorn==20.1.0
uvicorn==0.19.0
protobuf==3.19.1
